Refactor to move all top-level code to appropriate places
This refactoring extracts the program logic of tuning for Timescale
into a package called tstune in pkg/tstune while moving the command
line binary program under cmd/. Testing is easier with this
structure since you can now pass any io.Reader/Writer into the
created tstune.Tuner object to test input/output paths.
