+++ pkg/tstune/checker_test.go
@@ -1,4 +1,4 @@
-package main
+package tstune
 
 import "testing"
